Updated the proto file to contain the API for our chat application. Renamed it from HelloWorld to GRPCWebChat.
