+++ backend/pom.xml
@@ -6,7 +6,7 @@
 
     <parent>
         <groupId>com.angelomelonas</groupId>
-        <artifactId>grpcwebhelloworld</artifactId>
+        <artifactId>grpcwebchat</artifactId>
         <version>0.0.1</version>
     </parent>
