+++ pom.xml
@@ -5,7 +5,7 @@
     <modelVersion>4.0.0</modelVersion>
 
     <groupId>com.angelomelonas</groupId>
-    <artifactId>grpcwebhelloworld</artifactId>
+    <artifactId>grpcwebchat</artifactId>
     <version>0.0.1</version>
     <packaging>pom</packaging>
