--- proto/helloworld.proto
@@ -1,40 +1,0 @@
-// Copyright 2018 Google LLC
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     https://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-syntax = "proto3";
-
-package helloworld;
-option java_package = "com.angelomelonas.helloworld";
-
-service Greeter {
-  // unary call
-  rpc SayHello(HelloRequest) returns (HelloReply);
-  // server streaming call
-  rpc SayRepeatHello(RepeatHelloRequest) returns (stream HelloReply);
-  // unary call - response after a length delay
-  rpc SayHelloAfterDelay(HelloRequest) returns (HelloReply);
-}
-
-message HelloRequest {
-  string name = 1;
-}
-
-message RepeatHelloRequest {
-  string name = 1;
-  int32 count = 2;
-}
-
-message HelloReply {
-  string message = 1;
-}
